% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalCorrHuman.R
\name{AnalCorrHuman}
\alias{AnalCorrHuman}
\title{AnalCorrHuman}
\usage{
AnalCorrHuman(DfList_Human_ProteinCorrelations, pValue_Threshold = 0.05)
}
\arguments{
\item{DfList_Human_ProteinCorrelations}{Output from 'MakeCorrHuman' with parameter 'runAnalCorrHuman=FALSE'}

\item{pValue_Threshold}{Protein correlations with a pValue smaller than this threshold are considered significant, Default: 0.05}
}
\value{
Data frame with all human proteins subjected to correlations and annotated significance across donors
}
\description{
Function to extract proteins which significanlty correlate with MHCI peptide counts across tissues and donors.
}
\details{
This function should always be used as part of the function 'MakeCorrHuman', only to be run seperately to save on computation time/space
}
\examples{
\dontrun{
 df_corrhuman <- AnalCorrHuman(MakeCorrHumanRaw,pValue_Threshold = 0.05)
}
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_path}},\code{\link[ggplot2]{scale_colour_brewer}},\code{\link[ggplot2]{geom_abline}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{facet_grid}},\code{\link[ggplot2]{lims}}
}
