% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicAnalHuman.R
\name{BasicAnalHuman}
\alias{BasicAnalHuman}
\title{BasicAnalHuman}
\usage{
BasicAnalHuman(df_human, df_mouse = NULL)
}
\arguments{
\item{df_human}{Human MHCI data frame, usually an output from the function 'GetHumanMHCIdata'}

\item{df_mouse}{Optional parameter; Mouse MHCI data frame, usually an output from the function 'GetMouseMHCIdata', Default: NULL}
}
\value{
List of plots of the analysis
}
\description{
Function to provide basic analysis and visualization of human MHCI tissue draft (MHCI peptides per tissue and principal component analysis)
}
\details{
If df_mouse is specified, mouse and human MHCI tissue drafts are directly compared
}
\examples{
\dontrun{
BasicAnalHuman(df_human)
}
}
\seealso{
\code{\link[dplyr]{tally}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise_all}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_boxplot}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{geom_smooth}},\code{\link[ggplot2]{geom_point}}
 \code{\link[forcats]{fct_reorder}}
 \code{\link[FactoMineR]{PCA}}
 \code{\link[factoextra]{get_pca}},\code{\link[factoextra]{fviz_pca}}
 \code{\link[ggpmisc]{stat_poly_eq}}
 \code{\link[ggrepel]{geom_label_repel}}
 \code{\link[cowplot]{ggdraw}},\code{\link[cowplot]{draw_plot}},\code{\link[cowplot]{draw_plot_label}}
}
