% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicAnalMouse.R
\name{BasicAnalMouse}
\alias{BasicAnalMouse}
\title{BasicAnalMouse}
\usage{
BasicAnalMouse(df_mouse, returnPlotsOnly = FALSE)
}
\arguments{
\item{df_mouse}{Mouse MHCI data frame, usually an output from the function 'GetMouseMHCIdata'}

\item{returnPlotsOnly}{Only plots will be returned if TRUE and plots and data will be returned as a list if set to FALSE, Default: FALSE}
}
\value{
List of plots and data
}
\description{
Function to provide basic analysis and visualization of mouse MHCI tissue draft (MHCI peptides per tissue and principal component analysis)
}
\details{
Returned data frame is a matrix of peptide intensities across tissues
}
\examples{
\dontrun{
BasicAnalMouse(df_mouse)
}
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_bar}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{scale_colour_brewer}}
 \code{\link[FactoMineR]{PCA}}
 \code{\link[factoextra]{get_pca}},\code{\link[factoextra]{fviz_pca}}
 \code{\link[cowplot]{ggdraw}},\code{\link[cowplot]{draw_plot}},\code{\link[cowplot]{draw_plot_label}}
}
