% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparePeaksHuman.R
\name{ComparePeaksHuman}
\alias{ComparePeaksHuman}
\title{ComparePeaksHuman}
\usage{
ComparePeaksHuman(peaks_peptide_output, df_human)
}
\arguments{
\item{peaks_peptide_output}{The data frame containing the peaks output (peptide.csv)}

\item{df_human}{Data frame containing the human MHCI atlas data}
}
\value{
venn diagramm plot and dataframe of input annotated with presence/absence of these peptides in the human MHCI atlas
}
\description{
Function to compare human MHCI atlas to simple Peaks output (peptide.csv) to assess extent of overlapping peptides.
}
\examples{
\dontrun{
 ComparePeaksHuman(peaks_peptide_output,df_human)
}
}
\seealso{
\code{\link[limma]{venn}}
}
