% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConservationHuman.R
\name{ConservationHuman}
\alias{ConservationHuman}
\title{ConservationHuman}
\usage{
ConservationHuman(
  df_human,
  HumanHousekeepers,
  pathBW_human = "~/Downloads/hg38.phastCons100way.bw",
  samplesize = 2000,
  quantile = 4,
  ts_DonorSpecific = FALSE,
  MinTissuesPerDonor = 15,
  returnplots = TRUE
)
}
\arguments{
\item{df_human}{Human MHCI dataset, usually output from the function GetHumanMHCIdata}

\item{HumanHousekeepers}{List object from the function HousekeepersHuman}

\item{pathBW_human}{Path to bigwig file containing human PhastCons conservation scores, Default: '~/Downloads/hg38.phastCons100way.bw'}

\item{samplesize}{Number of tissue specific genes that will be randomly selected to compute conservation, Default: 2000}

\item{quantile}{Intensity quantile above which tissue specific peptides will be chosen (4 means that only the 25 percent most intense peptides are considered, 1 means that all peptides are considered), Default: 4}

\item{ts_DonorSpecific}{Logical value to set if tissue specific (ts) peptides are filtered donor by donor (TRUE) or across all donors (FALSE), Default: FALSE}

\item{MinTissuesPerDonor}{Peptide must be present in samples from a donor for which the specified minimum number of tissues was measured, Default: 15}

\item{returnplots}{Logical value to return or not return visualization object in output list, Default: TRUE}
}
\value{
List containing ggplot2 objects (plots) and output data to create plots
}
\description{
Function to calculate conservation rates of genes representing tissue specific MHCI peptides versus genes representing universal (found across all or most tissues/donors) MHCI peptides
}
\details{
Output data using the default values are provided with the package and can be accessed throug running the function 'mkFigure5'
}
\examples{
\dontrun{
cons_h<-ConservationHuman(df_human,HousekeepersHuman(df_human),samplesize=200)
}
}
\seealso{
\code{\link[TxDb.Hsapiens.UCSC.hg38.knownGene]{TxDb.Hsapiens.UCSC.hg38.knownGene}}
 \code{\link[dplyr]{count}}
 \code{\link[org.Hs.eg.db]{org.Hs.eg.db}}
 \code{\link[GenomicFeatures]{transcripts}},\code{\link[GenomicFeatures]{transcriptsBy}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{scale_colour_brewer}},\code{\link[ggplot2]{geom_path}},\code{\link[ggplot2]{facet_grid}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}},\code{\link[ggplot2]{labs}}
 \code{\link[cowplot]{ggdraw}},\code{\link[cowplot]{draw_plot}},\code{\link[cowplot]{draw_plot_label}}
 \code{\link[ggplotify]{as.ggplot}}
}
