% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConservationMouse.R
\name{ConservationMouse}
\alias{ConservationMouse}
\title{ConservationMouse}
\usage{
ConservationMouse(
  df_mouse,
  pathBW_mouse = "~/Downloads/mm10.60way.phastCons.bw",
  samplesize = 2000,
  returnplots = TRUE
)
}
\arguments{
\item{df_mouse}{Mouse MHCI dataset, usually output from the function 'GetMouseMHCIdata'}

\item{pathBW_mouse}{path to the BigWig file mm10.60way.phastCons.bw downloadable from https://genome.ucsc.edu/, Default: '~/Downloads/mm10.60way.phastCons.bw'}

\item{samplesize}{Number of tissue specific genes that will be randomly selected to compute conservation, Default: 2000}

\item{returnplots}{Logical value to return or not return visualization object in output list, Default: TRUE}
}
\value{
List containing ggplot2 objects (plots) and output data to create plots
}
\description{
Function to calculate conservation rates of genes representing tissue specific MHCI peptides versus genes representing universal (found across all or most tissues) MHCI peptides
}
\details{
Output data using the default values are provided with the package and can be accessed throug running the function 'mkFigure5'
}
\examples{
\dontrun{
cons_m<-ConservationMouse(df_mouse,samplesize = 200)
}
}
\seealso{
\code{\link[TxDb.Mmusculus.UCSC.mm10.knownGene]{TxDb.Mmusculus.UCSC.mm10.knownGene}}
 \code{\link[org.Mm.eg.db]{org.Mm.eg.db}}
 \code{\link[GenomicFeatures]{transcripts}},\code{\link[GenomicFeatures]{transcriptsBy}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{scale_colour_brewer}},\code{\link[ggplot2]{geom_path}},\code{\link[ggplot2]{facet_grid}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}},\code{\link[ggplot2]{labs}}
 \code{\link[cowplot]{ggdraw}}
 \code{\link[ggplotify]{as.ggplot}}
}
