% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccessions.R
\name{getAccessions}
\alias{getAccessions}
\title{getAccessions}
\usage{
getAccessions(x, y)
}
\arguments{
\item{x}{fasta header}

\item{y}{position of string separated by '|' that is to be eaxtracted (From the right starting at position 0)}
}
\value{
Substring
}
\description{
Function to extract entries from a UniprotKB/Swissprot fasta header
}
\details{
Used withing the package MHCIatlas to aid data analysis
}
\examples{
\dontrun{
getAccessions('sp|P55012|S12A2_MOUSE',0)
> "S12A2"
}
}
