% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHumanMHCIdata.R
\name{GetHumanMHCIdata}
\alias{GetHumanMHCIdata}
\title{GetHumanMHCIdata}
\usage{
GetHumanMHCIdata(NetMHC_Rank_Threshold = 2, return_all_rawData = FALSE)
}
\arguments{
\item{NetMHC_Rank_Threshold}{NetMHCpan4.0 rank threshold value (peptides with binding scores below or equal to this value are selected), Default: 2}

\item{return_all_rawData}{Logical, TRUE returns the raw dataset while FALSE returns a cleaned up version that is to be used with most functions of the MHCIatlas Rpackage, Default: FALSE}
}
\value{
Data frame of the human MHCI peptide tissue draft
}
\description{
Function to retrieve the human MHCI tissue draft dataset
}
\details{
More about the NetMHCpan-4.0 scoring system and MHCI binding prediction can be found here:
    NetMHCpan-4.0: Improved Peptide–MHC Class I Interaction Predictions Integrating Eluted Ligand and Peptide
    Binding Affinity Data
    Vanessa Jurtz, Sinu Paul, Massimo Andreatta, Paolo Marcatili, Bjoern Peters and Morten Nielsen
    The Journal of Immunology (2017) ji1700893; DOI: 10.4049/jimmunol.1700893
}
\examples{
\dontrun{
df_human<- GetHumanMHCIdata()
}
}
