% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HousekeepersHuman.R
\name{HousekeepersHuman}
\alias{HousekeepersHuman}
\title{HousekeepersHuman}
\usage{
HousekeepersHuman(
  df_human,
  NumDonorsperGene = 2,
  NumAllelesperGene = 1,
  topSeq = 100,
  minNumTissues = 14
)
}
\arguments{
\item{df_human}{Input data frame usually from GetHumanMHCI data}

\item{NumDonorsperGene}{Minimum number of donors where a gene was found to be represented by universal (housekeeping) MHCI peptides, Default: 2}

\item{NumAllelesperGene}{Minimum number of alleles across all donors where a gene was found to be represented by universal (housekeeping) MHCI peptides, Default: 1}

\item{topSeq}{Selection of the top x peptides measured the most frequently as univeral (housekeeping) from which genes will be determined as universal, Default: 100}

\item{minNumTissues}{Min number of tissues that must have been sampled for a donor to be considered, Default: 14}
}
\value{
List containing ggplot2 figures and output data frame
}
\description{
Functon to retrieve genes of MHCI peptides represented across most tissues/donors/alleles
}
\details{
The term universal peptide is used synonymously to housekeeping peptide
}
\examples{
\dontrun{
HousekeepersHuman(df_human)
}
}
\seealso{
\code{\link[dplyr]{count}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_bar}},\code{\link[ggplot2]{coord_flip}},\code{\link[ggplot2]{scale_continuous}}
 \code{\link[limma]{venn}}
 \code{\link[org.Hs.eg.db]{org.Hs.eg.db}}
}
