% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCorrHuman.R
\name{MakeCorrHuman}
\alias{MakeCorrHuman}
\title{MakeCorrHuman}
\usage{
MakeCorrHuman(
  df_human,
  Donors = c("all"),
  deconvolute_byHLAGene = FALSE,
  pValue_Threshold = 0.05,
  runAnalCorrHuman = TRUE
)
}
\arguments{
\item{df_human}{Input MHCI tissue draft from GetHumanMHCIdata}

\item{Donors}{Vector with names of donors for which correlations are to be computed, Default: c("all")}

\item{deconvolute_byHLAGene}{Logical, can be used to deconvolute by allele instead of by donor (not recommended), Default: FALSE}

\item{pValue_Threshold}{Maximum pValue at which a correlation will be considered significant, Default: 0.05}

\item{runAnalCorrHuman}{Logical, directly runs the function AnalCorrHuman to retrieve significantly correlating proteins, Default: TRUE}
}
\value{
List of dataframes with results
}
\description{
Compute correlations between tissue dependent intensities of proteins and the tissue dependent number of MHCI peptides
}
\details{
Running this function for all donors might take a while (around 15 minutes). Default pValue threshold is higher than for mouse since by default a pValue smaller the set value must be found in at least two donors/alleles
Human protein expression data were retrieved from: Wang et al.; Mol Syst Biol. 2019 Feb 18;15(2); PMID: 30777892 https://www.embopress.org/doi/full/10.15252/msb.20188503
}
\examples{
\dontrun{
corrs_h<-MakeCorrHuman(df_human,Donors=c('AUT01-DN11'))
}
}
\seealso{
\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}},\code{\link[dplyr]{setops}}
 \code{\link[stats]{lm}}
 \code{\link[broom]{reexports}}
}
