% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCorrMouse.R
\name{MakeCorrMouse}
\alias{MakeCorrMouse}
\title{MakeCorrMouse}
\usage{
MakeCorrMouse(df_mouse, pValue_Threshold = 0.01, useSILAC = F)
}
\arguments{
\item{df_mouse}{Input MHCI tissue draft from GetMouseMHCIdata}

\item{pValue_Threshold}{Maximum pValue at which a correlation will be considered significant, Default: 0.01}

\item{useSILAC}{Logical, use SILAC fold change values instead of normalized raw intensities of protein expression data, Default: F}
}
\value{
List of dataframes with results
}
\description{
Compute correlations between tissue dependent intensities of proteins and the tissue dependent number of MHCI peptides
}
\details{
Mouse protein expression data are retireved from this publication: Geiger et al. Molecular & Cellular Proteomics June 1, 2013, 12 (6) 1709-1722; PMID: 30777892 https://www.mcponline.org/content/12/6/1709
}
\examples{
\dontrun{
corrs_m<-MakeCorrMouse(df_mouse)
}
}
\seealso{
\code{\link[dplyr]{tally}}
 \code{\link[stats]{lm}}
 \code{\link[broom]{reexports}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_path}},\code{\link[ggplot2]{scale_colour_brewer}},\code{\link[ggplot2]{geom_abline}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{facet_grid}},\code{\link[ggplot2]{lims}}
}
