% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkFigure3.R
\name{mkFigure3}
\alias{mkFigure3}
\title{mkFigure3}
\usage{
mkFigure3(df_human, df_mouse)
}
\arguments{
\item{df_human}{Human MHCI tissue draft data frame from GetHumanMHCI data}

\item{df_mouse}{Mouse MHCI tissue draft data frame from GetMouseMHCI data}
}
\value{
List with ggplot2 objects and data to generate the figure
}
\description{
Function to plot Figure 3 of the manuscript
}
\details{
DETAILS
}
\examples{
\dontrun{
F3 <- mkFigure3(df_human,df_mouse)
}
}
\seealso{
\code{\link[cowplot]{ggdraw}},\code{\link[cowplot]{draw_plot}},\code{\link[cowplot]{draw_plot_label}}
 \code{\link[ggplot2]{theme}}
 \code{\link[ggplotify]{as.ggplot}}
}
