% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkFigure4.R
\name{mkFigure4}
\alias{mkFigure4}
\title{mkFigure4}
\usage{
mkFigure4(df_mouse)
}
\arguments{
\item{df_mouse}{Mouse MHCI tissue draft data frame from GetMouseMHCI data}
}
\value{
List with ggplot2 objects and data to generate the figure
}
\description{
Function to plot Figure 4 of the manuscript
}
\details{
Mouse mRNA expression data were retrieved from: Söllner, J., Leparc, G., Hildebrandt, T. et al. An RNA-Seq atlas of gene expression in mouse and rat normal tissues. Sci Data 4, 170185 (2017). https://doi.org/10.1038/sdata.2017.185
}
\examples{
\dontrun{
F4 <- mkFigure4(df_human,df_mouse)
}
}
\seealso{
\code{\link[org.Mm.eg.db]{org.Mm.eg.db}}
 \code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{tally}}
 \code{\link[pheatmap]{pheatmap}}
 \code{\link[RColorBrewer]{RColorBrewer}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_violin}},\code{\link[ggplot2]{facet_wrap}},\code{\link[ggplot2]{geom_boxplot}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}},\code{\link[ggplot2]{scale_colour_brewer}}
 \code{\link[webr]{PieDonut}}
 \code{\link[cowplot]{ggdraw}},\code{\link[cowplot]{draw_plot}},\code{\link[cowplot]{draw_plot_label}}
 \code{\link[ggplotify]{as.ggplot}}
}
