% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkFigure5.R
\name{mkFigure5}
\alias{mkFigure5}
\title{mkFigure5}
\usage{
mkFigure5(
  df_human,
  HumanHousekeepers,
  df_mouse,
  useDefaultCons = TRUE,
  ConsMouse = NA,
  ConsHuman = NA
)
}
\arguments{
\item{df_human}{Human MHCI tissue draft data frame from GetHumanMHCI data}

\item{HumanHousekeepers}{List output from function HousekeepersHuman}

\item{df_mouse}{Mouse MHCI tissue draft data frame from GetMouseMHCI data}

\item{useDefaultCons}{Uses conservatio scores calculated with defulat values as in manuscript, Default: TRUE}

\item{ConsMouse}{Output list from ConservationMouse (Needed if useDefaultCons is set to FLASE), Default: NA}

\item{ConsHuman}{Output list from ConservationHuman (Needed if useDefaultCons is set to FLASE), Default: NA}
}
\value{
List with ggplot2 objects and data to generate the figure
}
\description{
Function to plot Figure 4 of the manuscript
}
\details{
DETAILS
}
\examples{
\dontrun{
F5<- mkFigure5(df_human,cons_h,df_mouse)
}
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{scale_colour_brewer}},\code{\link[ggplot2]{geom_path}},\code{\link[ggplot2]{facet_grid}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{geom_bar}},\code{\link[ggplot2]{lims}},\code{\link[ggplot2]{geom_boxplot}},\code{\link[ggplot2]{ggtheme}}
 \code{\link[cowplot]{ggdraw}},\code{\link[cowplot]{draw_plot}},\code{\link[cowplot]{draw_plot_label}}
 \code{\link[ggplotify]{as.ggplot}}
 \code{\link[org.Mm.eg.db]{org.Mm.eg.db}}
}
