% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkFigure6.R
\name{mkFigure6}
\alias{mkFigure6}
\title{mkFigure6}
\usage{
mkFigure6(corr_human, corr_mouse, GO_human = NA, GO_mouse = NA)
}
\arguments{
\item{corr_human}{correlation output from MakeCorrHuman}

\item{corr_mouse}{correlation output from MakeCorrMouse}

\item{GO_human}{data frame containing GO_terms for human proteins that significantly correlate (See output corr_human), file contains column with GO_terms named 'Gene.Set.Name' and column with qvalue named 'FDR.qvalue', Default is taking GO terms as in publication, Default: NA}

\item{GO_mouse}{data frame containing GO_terms for mouse proteins that significantly correlate (See output corr_human), file contains column with GO_terms named 'Gene.Set.Name' and column with qvalue named 'FDR.qvalue', Default is taking GO terms as in publication, Default: NA}
}
\value{
Figure 6 and GO term data
}
\description{
makes plot for Figure 6 B,C and D (A is only the cartoon)
}
\details{
DETAILS
}
\examples{
\dontrun{
mkFigure6(corr_human,corr_mouse)
}
}
\seealso{
\code{\link[reshape2]{melt}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_bar}},\code{\link[ggplot2]{coord_flip}},\code{\link[ggplot2]{scale_colour_gradient}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}},\code{\link[ggplot2]{scale_x_discrete}}
 \code{\link[scales]{label_wrap}}
 \code{\link[cowplot]{ggdraw}},\code{\link[cowplot]{draw_plot}},\code{\link[cowplot]{draw_plot_label}}
}
