% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkHumanConnectivityMap.R
\name{mkHumanConnectivityMap}
\alias{mkHumanConnectivityMap}
\title{mkHumanConnectivityMap}
\usage{
mkHumanConnectivityMap(
  df_human,
  returnPlotsOnly = FALSE,
  AlleleEnrichThr = 1.5
)
}
\arguments{
\item{df_human}{Data frame containing MHCI tissue draft data from GetHumanMHCIdata}

\item{returnPlotsOnly}{Loical, Default: FALSE}

\item{AlleleEnrichThr}{Min fold change value above which an allele is considered enriched within a given tissue, Default: 1.5}
}
\value{
list with ggplot2 objects and results data
}
\description{
Function to generate human connectivity map and allele enrichment map
}
\examples{
\dontrun{
mkHumanConnectivityMap(df_human, returnPlotsOnly=F,AlleleEnrichThr=1.5)
}
}
\seealso{
\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}},\code{\link[dplyr]{n}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{tally}},\code{\link[dplyr]{mutate_all}}
 \code{\link[tidyr]{spread}}
 \code{\link[reshape2]{melt}},\code{\link[reshape2]{cast}}
 \code{\link[RColorBrewer]{RColorBrewer}}
 \code{\link[pheatmap]{pheatmap}}
 \code{\link[Matrix]{forceSymmetric}}
}
