% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkMouseConnectivityMap.R
\name{mkMouseConnectivityMap}
\alias{mkMouseConnectivityMap}
\title{mkMouseConnectivityMap}
\usage{
mkMouseConnectivityMap(df_mouse)
}
\arguments{
\item{df_mouse}{Data frame containing MHCI tissue draft data from GetMouseMHCIdata}
}
\value{
list with ggplot2 objects and results data
}
\description{
unction to generate mouse connectivity map
}
\examples{
\dontrun{
mkMouseConnectivityMap(df_mouse)
}
}
\seealso{
\code{\link[reshape2]{melt}},\code{\link[reshape2]{cast}}
 \code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}},\code{\link[dplyr]{n}},\code{\link[dplyr]{mutate_all}}
 \code{\link[Matrix]{forceSymmetric}}
 \code{\link[pheatmap]{pheatmap}}
 \code{\link[RColorBrewer]{RColorBrewer}}
}
