% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsHumanProtCorr.R
\name{PlotsHumanProtCorr}
\alias{PlotsHumanProtCorr}
\title{PlotsHumanProtCorr}
\usage{
PlotsHumanProtCorr(
  Human_ProtCorrelations,
  SigGene_names = NULL,
  allSigprots = TRUE,
  RankSigThreshold = 2,
  path_filename = NA,
  return_list_of_plots = TRUE
)
}
\arguments{
\item{Human_ProtCorrelations}{List which is output object from the function MakeCorrHuman}

\item{SigGene_names}{List of genes of interest, Default: NULL}

\item{allSigprots}{Logical to generate plots for all genes (TRUE) or for a specified list of genes (FALSE), Default: TRUE}

\item{RankSigThreshold}{Number of donors where a gene has to be found with significant correlation, Default: 2}

\item{path_filename}{path and filename where a pdf file containing all plots is to be written to (optional), Default: NA}

\item{return_list_of_plots}{Logical, can return a list with plots for each gene, Default: TRUE}
}
\value{
list of plots for each gene as ggplot2 object
}
\description{
Function to visualize correlations of protein intensities and MHCI peptide counts across tissues and donors in human
}
\examples{
\dontrun{
plots_h<-PlotsHumanProtCorr(corrs_h,SigGene_names=c('HLA-B'),
allSigprots=F,path_filename='~/HLAB.pdf')
}
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_smooth}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{scale_colour_brewer}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}}
 \code{\link[ggpmisc]{stat_poly_eq}}
 \code{\link[ggrepel]{geom_label_repel}}
}
