% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsMouseProtCorr.R
\name{PlotsMouseProtCorr}
\alias{PlotsMouseProtCorr}
\title{PlotsMouseProtCorr}
\usage{
PlotsMouseProtCorr(
  Mouse_ProtCorrelations,
  pValue_Threshold = 0.01,
  SigGene_names = NULL,
  path_filename = NA,
  return_list_of_plots = TRUE
)
}
\arguments{
\item{Mouse_ProtCorrelations}{List which is output object from the function MakeCorrMouse}

\item{pValue_Threshold}{Max pValue below which correlations are considered signifcant, Default: 0.01}

\item{SigGene_names}{List of genes of interest, Default: NULL, Default: NULL}

\item{path_filename}{path and filename where a pdf file containing all plots is to be written to (optional), Default: NA}

\item{return_list_of_plots}{Logical, can return a list with plots for each gene, Default: TRUE}
}
\value{
list of plots for each gene as ggplot2 object
}
\description{
Function to visualize correlations of protein intensities and MHCI peptide counts across tissues in mouse
}
\details{
ggplot2 plots can directly be accessed from the list of plots
}
\examples{
\dontrun{
plots_m<-PlotsMouseProtCorr(corrs_m,SigGene_names=c('Erap1','Psmb1'),
allSigprots=F,path_filename='~/Erap1andPsmb1.pdf')
}
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_smooth}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{scale_colour_brewer}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}}
 \code{\link[ggpmisc]{stat_poly_eq}}
 \code{\link[ggrepel]{geom_label_repel}}
}
