% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProgressBar.R
\name{ProgressBar}
\alias{ProgressBar}
\title{ProgressBar}
\usage{
ProgressBar(i = 1, len_i = 100, j = 1, len_j = 1)
}
\arguments{
\item{i}{integer i, Default: 1}

\item{len_i}{maximum value of i, Default: 100}

\item{j}{integer j, Default: 1}

\item{len_j}{maximum value of j, Default: 1}
}
\value{
A progress bar
}
\description{
Draws progress bar, useful to show progress in loops and nested loops
}
\examples{
\dontrun{
ProgressBar(1,3)
}
}
