% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigCorrsHuman.R
\name{SigCorrsHuman}
\alias{SigCorrsHuman}
\title{SigCorrsHuman}
\usage{
SigCorrsHuman(Human_ProtCorrelations, RankSig_Threshold = 2)
}
\arguments{
\item{Human_ProtCorrelations}{Output list from function MakeCorrHuman}

\item{RankSig_Threshold}{Min number of occurrences where correlation was found significant, Default: 2}
}
\value{
Data frame containing significantly correlating genes and the mean correlation pValue
}
\description{
Retrieve significantly correlating proteins and their gene names
}
\details{
Used to get an overview of singificantly proteins that can be used for later gene ontology analysis etc.
}
\examples{
\dontrun{
sig_prots_h<-SigCorrsHuman(corrs_h)
}
}
