% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigCorrsMouse.R
\name{SigCorrsMouse}
\alias{SigCorrsMouse}
\title{SigCorrsMouse}
\usage{
SigCorrsMouse(Mouse_ProtCorrelations, pValue_Threshold = 0.01)
}
\arguments{
\item{Mouse_ProtCorrelations}{Output list from function MakeCorrMouse}

\item{pValue_Threshold}{maximum pValue at or below which a correlation is considered significant, Default: 0.01}
}
\value{
Data frame containing significantly correlating genes and the mean correlation pValue
}
\description{
Retrieve significantly correlating proteins and their gene names
}
\details{
Used to get an overview of singificantly proteins that can be used for later gene ontology analysis etc.
}
\examples{
\dontrun{
sig_prots_m<-SigCorrsHuman(corrs_m)
}
}
