% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupplFigure10.R
\name{SupplFigure10}
\alias{SupplFigure10}
\title{SupplFigure10}
\usage{
SupplFigure10(corr_human, corr_mouse, plots_human, plots_mouse)
}
\arguments{
\item{corr_human}{Correlation analysis human from MakeCorrHuman}

\item{corr_mouse}{Correlation analysis mouse from MakeCorrMouse}

\item{plots_human}{Plots human from PlotsHumanProtCorr}

\item{plots_mouse}{Plots mouse from PlotsMouseProtCorr}
}
\value{
Supplementary Figure 10
}
\description{
Plot supplementary figure 10
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_bar}},\code{\link[ggplot2]{scale_colour_brewer}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}}
 \code{\link[cowplot]{ggdraw}},\code{\link[cowplot]{draw_plot}},\code{\link[cowplot]{draw_plot_label}}
}
