% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupplFigure2.R
\name{SupplFigure2}
\alias{SupplFigure2}
\title{SupplFigure2}
\usage{
SupplFigure2(df_human)
}
\arguments{
\item{df_human}{Human immonopeptidome data frame from function GetHumanMHCIdata}
}
\value{
Supplementary Figure 2 as plot
}
\description{
Function to plot Supplementary Figure 2
}
\examples{
\dontrun{
SupplFigure2(df_human)
}
}
\seealso{
\code{\link[data.table]{as.data.table}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_bar}},\code{\link[ggplot2]{coord_flip}},\code{\link[ggplot2]{labs}}
 \code{\link[cowplot]{ggdraw}},\code{\link[cowplot]{draw_plot}},\code{\link[cowplot]{draw_plot_label}}
}
