% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupplFigure3.R
\name{SupplFigure3}
\alias{SupplFigure3}
\title{SupplFigure3}
\usage{
SupplFigure3(df_mouse)
}
\arguments{
\item{df_mouse}{Mouse immunopeptidomicvs dataframe from GetMouseMHCIdata}
}
\value{
supplementary figure 3
}
\description{
Plot supplementary figure 3
}
\seealso{
\code{\link[stats]{reshape}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_smooth}},\code{\link[ggplot2]{scale_continuous}},\code{\link[ggplot2]{sec_axis}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}}
}
