% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupplFigure4.R
\name{SupplFigure4}
\alias{SupplFigure4}
\title{SupplFigure4}
\usage{
SupplFigure4(df_human)
}
\arguments{
\item{df_human}{Human immunopeptidomicvs dataframe from GetHumanMHCIdata}
}
\value{
supplementary figure 4
}
\description{
Plot supplementary figure 4
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_smooth}},\code{\link[ggplot2]{facet_wrap}}
}
