% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupplFigure5.R
\name{SupplFigure5}
\alias{SupplFigure5}
\title{SupplFigure5}
\usage{
SupplFigure5(df_human)
}
\arguments{
\item{df_human}{Human immunopeptidomicvs dataframe from GetHumanMHCIdata}
}
\value{
supplementary figure 5
}
\description{
Plot supplementary figure 5
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_smooth}},\code{\link[ggplot2]{facet_wrap}}
}
