% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupplFigure6.R
\name{SupplFigure6}
\alias{SupplFigure6}
\title{SupplFigure6}
\usage{
SupplFigure6(df_mouse, df_human)
}
\arguments{
\item{df_mouse}{Mouse immunopeptidomicvs dataframe from GetMouseMHCIdata}

\item{df_human}{Human immunopeptidomicvs dataframe from GetHumanMHCIdata}
}
\value{
supplementary figure 6
}
\description{
Plot supplementary figure 6
}
\seealso{
\code{\link[dplyr]{count}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_bar}},\code{\link[ggplot2]{labs}}
 \code{\link[ggforce]{facet_zoom}}
 \code{\link[cowplot]{ggdraw}},\code{\link[cowplot]{draw_plot}},\code{\link[cowplot]{draw_plot_label}}
}
